feature(calculator) : add Integer parser class
